MP4Splitter: добавлена поддержка пропорций кадра для HEVC.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@3548 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/DSUtil/VideoParser.h
@@ -76,7 +76,7 @@ void CreateSequenceHeaderHEVC(BYTE* data, int size, DWORD* dwSequenceHeader, DWO
 ////
 
 struct vc_params_t {
-	DWORD width, height;
+	LONG width, height;
 
 	DWORD profile, level;
 	DWORD nal_length_size;
